Add basic support for iOS 26 relocation
Expose `tcc_relocate_ex` and tweak the logic so that `ptr_diff`
can be used for MACHO targets to express a rigid offset from the
writable memory passed in `ptr` to the actual executable address used
for relocation.

In this way the compiled executable is copied to the writable view of
the memory, `mprotect` acts on the already-executable ranges having no
negative impact, and the code is retro-compatible because calling
`tcc_relocate` will have `ptr_diff` set to `0` (except on selinux when
internal allocation is enabled) behaving exactly as before minimizing
the impact on the library size and complexity.